[core] Make ManifestEntry and DataFileMeta to interfaces
