+++ paimon-core/src/main/java/org/apache/paimon/manifest/ManifestEntry.java
@@ -26,12 +26,9 @@ import org.apache.paimon.types.IntType;
 import org.apache.paimon.types.RowType;
 import org.apache.paimon.types.TinyIntType;
 
-import javax.annotation.Nullable;
-
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 
 import static org.apache.paimon.utils.SerializationUtils.newBytesType;
 
@@ -41,9 +38,9 @@ import static org.apache.paimon.utils.SerializationUtils.newBytesType;
  * @since 0.9.0
  */
 @Public
-public class ManifestEntry implements FileEntry {
+public interface ManifestEntry extends FileEntry {
 
-    public static final RowType SCHEMA =
+    RowType SCHEMA =
             new RowType(
                     false,
                     Arrays.asList(
@@ -53,158 +50,36 @@ public class ManifestEntry implements FileEntry {
                             new DataField(3, "_TOTAL_BUCKETS", new IntType(false)),
                             new DataField(4, "_FILE", DataFileMeta.SCHEMA)));
 
-    private final FileKind kind;
-    // for tables without partition this field should be a row with 0 columns (not null)
-    private final BinaryRow partition;
-    private final int bucket;
-    private final int totalBuckets;
-    private final DataFileMeta file;
-
-    public ManifestEntry(
+    static ManifestEntry create(
             FileKind kind, BinaryRow partition, int bucket, int totalBuckets, DataFileMeta file) {
-        this.kind = kind;
-        this.partition = partition;
-        this.bucket = bucket;
-        this.totalBuckets = totalBuckets;
-        this.file = file;
+        return new PojoManifestEntry(kind, partition, bucket, totalBuckets, file);
     }
 
-    @Override
-    public FileKind kind() {
-        return kind;
-    }
+    DataFileMeta file();
 
-    @Override
-    public BinaryRow partition() {
-        return partition;
-    }
+    ManifestEntry copyWithoutStats();
 
-    @Override
-    public int bucket() {
-        return bucket;
-    }
+    ManifestEntry assignSequenceNumber(long minSequenceNumber, long maxSequenceNumber);
 
-    @Override
-    public int level() {
-        return file.level();
-    }
+    ManifestEntry assignFirstRowId(long firstRowId);
 
-    @Override
-    public String fileName() {
-        return file.fileName();
-    }
+    byte[] toBytes() throws IOException;
 
-    @Nullable
-    @Override
-    public String externalPath() {
-        return file.externalPath().orElse(null);
-    }
-
-    @Override
-    public BinaryRow minKey() {
-        return file.minKey();
-    }
-
-    @Override
-    public BinaryRow maxKey() {
-        return file.maxKey();
-    }
-
-    @Override
-    public List<String> extraFiles() {
-        return file.extraFiles();
-    }
-
-    @Override
-    public int totalBuckets() {
-        return totalBuckets;
-    }
-
-    public DataFileMeta file() {
-        return file;
-    }
-
-    @Override
-    public Identifier identifier() {
-        return new Identifier(
-                partition,
-                bucket,
-                file.level(),
-                file.fileName(),
-                file.extraFiles(),
-                file.embeddedIndex(),
-                externalPath());
-    }
-
-    public ManifestEntry copyWithoutStats() {
-        return new ManifestEntry(kind, partition, bucket, totalBuckets, file.copyWithoutStats());
-    }
-
-    public ManifestEntry assignSequenceNumber(long minSequenceNumber, long maxSequenceNumber) {
-        return new ManifestEntry(
-                kind,
-                partition,
-                bucket,
-                totalBuckets,
-                file.assignSequenceNumber(minSequenceNumber, maxSequenceNumber));
-    }
-
-    public ManifestEntry assignFirstRowId(long firstRowId) {
-        return new ManifestEntry(
-                kind, partition, bucket, totalBuckets, file.assignFirstRowId(firstRowId));
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (!(o instanceof ManifestEntry)) {
-            return false;
-        }
-        ManifestEntry that = (ManifestEntry) o;
-        return Objects.equals(kind, that.kind)
-                && Objects.equals(partition, that.partition)
-                && bucket == that.bucket
-                && totalBuckets == that.totalBuckets
-                && Objects.equals(file, that.file);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(kind, partition, bucket, totalBuckets, file);
-    }
-
-    @Override
-    public String toString() {
-        return String.format("{%s, %s, %d, %d, %s}", kind, partition, bucket, totalBuckets, file);
-    }
-
-    public static long recordCount(List<ManifestEntry> manifestEntries) {
+    static long recordCount(List<ManifestEntry> manifestEntries) {
         return manifestEntries.stream().mapToLong(manifest -> manifest.file().rowCount()).sum();
     }
 
-    public static long recordCountAdd(List<ManifestEntry> manifestEntries) {
+    static long recordCountAdd(List<ManifestEntry> manifestEntries) {
         return manifestEntries.stream()
                 .filter(manifestEntry -> FileKind.ADD.equals(manifestEntry.kind()))
                 .mapToLong(manifest -> manifest.file().rowCount())
                 .sum();
     }
 
-    public static long recordCountDelete(List<ManifestEntry> manifestEntries) {
+    static long recordCountDelete(List<ManifestEntry> manifestEntries) {
         return manifestEntries.stream()
                 .filter(manifestEntry -> FileKind.DELETE.equals(manifestEntry.kind()))
                 .mapToLong(manifest -> manifest.file().rowCount())
                 .sum();
-    }
-
-    // ----------------------- Serialization -----------------------------
-
-    private static final ThreadLocal<ManifestEntrySerializer> SERIALIZER_THREAD_LOCAL =
-            ThreadLocal.withInitial(ManifestEntrySerializer::new);
-
-    public byte[] toBytes() throws IOException {
-        return SERIALIZER_THREAD_LOCAL.get().serializeToBytes(this);
-    }
-
-    public ManifestEntry fromBytes(byte[] bytes) throws IOException {
-        return SERIALIZER_THREAD_LOCAL.get().deserializeFromBytes(bytes);
     }
 }
