+++ paimon-core/src/main/java/org/apache/paimon/table/sink/CommitMessageLegacyV2Serializer.java
@@ -139,7 +139,7 @@ public class CommitMessageLegacyV2Serializer {
 
         @Override
         public DataFileMeta fromRow(InternalRow row) {
-            return new DataFileMeta(
+            return DataFileMeta.create(
                     row.getString(0).toString(),
                     row.getLong(1),
                     row.getLong(2),
