+++ paimon-core/src/test/java/org/apache/paimon/manifest/ManifestCommittableSerializerCompatibilityTest.java
@@ -58,7 +58,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -135,7 +135,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -211,7 +211,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -285,7 +285,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -359,7 +359,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -432,7 +432,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -506,7 +506,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -580,7 +580,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
@@ -654,7 +654,7 @@ public class ManifestCommittableSerializerCompatibilityTest {
                         singleColumn("max_value"),
                         fromLongArray(new Long[] {0L}));
         DataFileMeta dataFile =
-                new DataFileMeta(
+                DataFileMeta.create(
                         "my_file",
                         1024 * 1024,
                         1024,
